Move Nightscout api to new project
