+++ nightscout-api/src/main/java/com/dongtronic/nightscout/exceptions/NoNightscoutDataException.kt
@@ -1,0 +1,3 @@
+package com.dongtronic.nightscout.exceptions
+
+class NoNightscoutDataException : Exception()
