--- src/main/java/com/dongtronic/diabot/exceptions/NoNightscoutDataException.kt
@@ -1,3 +1,0 @@
-package com.dongtronic.diabot.exceptions
-
-class NoNightscoutDataException : Exception()
